+++ SpotBot/src/services/configuration/config_message_helpers.ts
