+++ SpotBot/src/services/configuration/configuration_service.ts
@@ -1,8 +1,8 @@
 import { Guild, TextChannel, Client, Message, ChannelType } from 'discord.js';
 import { IChannel, IConfig } from '../../interfaces/IConfig';
-import * as ConfigMessageHelpers from './configMessageHelpers';
+import * as ConfigMessageHelpers from './config_message_helpers';
 import { LogService } from '../log_service';
-import { autoInjectable, singleton } from 'tsyringe';
+import { singleton } from 'tsyringe';
 import { FileService } from '../file_service';
 import { WelcomeService } from '../welcome/welcome_service';
 const fs = require('fs');
