+++ SpotBot/src/services/help/help_service.ts
